+++ src/ContextFactory.php
@@ -31,10 +31,11 @@ class ContextFactory implements ContextFactoryInterface
             BodyParams::class => fn () => BodyParams::fromRequest($request),
             ServerParams::class => fn () => ServerParams::fromRequest($request),
             PathParams::class => fn () => PathParams::fromRequest($request),
-            Respond::class => fn ($c) => new Respond(
-                statusCode: $request->getMethod() === 'POST' ? 201 : 200,
-                contentType: $c->get('substance.http.default-content-type'),
-            ),
+            Respond::class => function ($c) use ($request) {
+                $respond = new Respond($request->getMethod() === 'POST' ? 201 : 200);
+                $respond->setHeader('Content-Type', $c->get('substance.http.default-content-type'));
+                return $respond;
+            },
         ];
     }
 }